+++ app/Services/EbookService.php
@@ -8,6 +8,7 @@ use App\Models\Product;
 use App\Repositories\Ebook\EbookRepositoryInterface;
 use App\Repositories\PaginationInterface;
 use Illuminate\Http\UploadedFile;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 
 class EbookService
@@ -62,7 +63,7 @@ class EbookService
             $dto->file = $uploadedFilePath;
         }
 
-        // Criar o produto
+        // Criar o produto utilizando o repositório
         return $this->repository->new($dto);
     }
 
@@ -73,30 +74,60 @@ class EbookService
 
     public function update(UpdateEbookDTO $dto): ?Product
     {
-        // Buscar a lição existente
         $ebook = $this->repository->findById($dto->id);
 
-        // Verificar instância da classe UploadedFile
-        if ($dto->image instanceof UploadedFile) {
-            if ($ebook && $ebook->image) {
-                // Remover o ficheiro existente, se houver
-                $this->uploadFile->removeFile($ebook->image);
-            }
-            // Processar o novo ficheiro
-            $file = $dto->image;
-            $customImageName = Str::of($dto->name)->slug('-') . '.' . $file->getClientOriginalExtension();
-            // Armazenar o novo ficheiro e obter o caminho do ficheiro armazenado
-            $uploadedFilePath = $this->uploadFile->storeAs($dto->image, 'lessonPdf', $customImageName);
+        if ($ebook && Gate::allows('owner-ebook', $ebook)) {
+            // Atualizar imagem se uma nova foi enviada
 
-            // Atualizar o DTO com o caminho relativo do ficheiro armazenado
-            $dto->image = $uploadedFilePath;
-        } else {
-            // Manter o caminho do ficheiro existente, se não houver novo ficheiro
-            unset($dto->image);
+            if ($dto->image instanceof UploadedFile) {
+                if ($ebook && $ebook->image) {
+                    // Remover o ficheiro existente, se houver
+                    $this->uploadFile->removeFile($ebook->image);
+                }
+
+                // Processar o novo ficheiro
+                $image = $dto->image;
+                $customImageName = Str::of($dto->code)->slug('-') . '.' . $image->getClientOriginalExtension();
+
+                // Armazenar o novo ficheiro e obter o caminho do ficheiro armazenado
+                $uploadedFilePath = $this->uploadFile->storeAs($dto->image, 'Products/ImagesEbooks', $customImageName);
+
+                // Atualizar o DTO com o caminho relativo do ficheiro armazenado
+                $dto->image = $uploadedFilePath;
+            } else {
+                // Manter o caminho do ficheiro existente, se não houver novo ficheiro
+                unset($dto->image);
+            }
+
+            if ($dto->file instanceof UploadedFile) {
+                if ($ebook && $ebook->file) {
+                    // Remover o ficheiro existente, se houver
+                    $this->uploadFile->removeFile($ebook->file);
+                }
+
+                // Processar o novo ficheiro
+                $file = $dto->file;
+                $customFileName = Str::of($dto->code)->slug('-') . '.' . $file->getClientOriginalExtension();
+
+                // Armazenar o novo ficheiro e obter o caminho do ficheiro armazenado
+                $uploadedFilePath = $this->uploadFile->storeAs($dto->file, 'Products/FilesEbooks', $customFileName);
+
+                // Atualizar o DTO com o caminho relativo do ficheiro armazenado
+                $dto->file = $uploadedFilePath;
+            } else {
+                // Manter o caminho do ficheiro existente, se não houver novo ficheiro
+                unset($dto->file);
+            }
+
+
+            // Atualizar o ebook com as novas informações
+            return $this->repository->update($dto);
         }
-        // Atualizar a lição no repositório com os dados do DTO
-        return $this->repository->update($dto);
+
+        return null;
     }
+
+
 
     public function delete(string $id)
     {
